+++ src/feature-toggle-service.ts
@@ -1,15 +1,25 @@
-let settings: { [k: string]: boolean } = {};
+const version = '4.0.1';
+
+let settings: {
+  [v: string]: { [k: string]: boolean }
+} = { [version]: {} };
 
 /**
  * Checks if Feature toggle/flag exists and it's on.
  *
  * @param key - Feature toggle/flag key to be checked
+ * @param debug - Prints configuration, version and settings for debug purposes
  *
  * @returns Boolean
  *
  */
-const isOn = (key: string): boolean => {
-  return !!settings[key];
+const isOn = (key: string, debug: boolean): boolean => {
+  if (debug === true) {
+    console.log(`
+[DEBUG] Execution Time in milliseconds: ${Date.now()} version: "${version}" key: "${key}" settings: ${JSON.stringify(settings)}
+`);
+  }
+  return !!settings[version][key];
 };
 
 /**
@@ -21,7 +31,7 @@ const isOn = (key: string): boolean => {
  *
  */
 const set = (obj: { [k: string]: boolean }): void => {
-  settings = obj;
+  settings[version] = obj;
 };
 
 export { isOn, set };